Cast LogLevel to int explicitly before use as int

+++ src/dpl/log/src/old_style_log_provider.cpp
@@ -95,7 +95,7 @@ void OldStyleLogProvider::Log(AbstractLogProvider::LogLevel level,
 			std::string("] ") << function << std::string("(): ") << message << mark.end;
 		fprintf(stdout, "%s\n", val.str().c_str());
 	} catch (const std::out_of_range &) {
-		fprintf(stdout, "Unsupported log level: %d\n", level);
+		fprintf(stdout, "Unsupported log level: %d\n", static_cast<int>(level));
 	}
 }
